Fixed!
It turns out completing doing away with the synchronisation issues caused by trying interrupt the DMA output for vsync, is the solution. So now on the ESP32S2 implementation, any call to mbi_set_pixel will be immediately reflected on screen. Tearing and all. This seems to work OK. Equivalent of not having double-buffering.